+++ src/parse.cpp
@@ -2,8 +2,6 @@
 #include "parse.h"
 
 Handle<Value> ParseField(const google::protobuf::Message &message, const Reflection *r, const FieldDescriptor *field, int index) {
-	NanScope();
-
 	Handle<Value> v;
 
 	switch (field->cpp_type()) {
@@ -13,7 +11,7 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				value = r->GetRepeatedInt32(message, field, index);
 			else
 				value = r->GetInt32(message, field);
-			v = Number::New(value);
+			v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_INT64: {
@@ -29,12 +27,12 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				uint32 hi, lo;
 				hi = (uint32) ((uint64)value >> 32);
 				lo = (uint32) value;
-				Local<Array> t = Array::New(2);
-				t->Set(0, Number::New(hi));
-				t->Set(1, Number::New(lo));
+				Local<Array> t = NanNew<Array>(2);
+				t->Set(0, NanNew<Number>(hi));
+				t->Set(1, NanNew<Number>(lo));
 				v = t;
 			} else
-				v = Number::New(value);
+				v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_UINT32: {
@@ -43,7 +41,7 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				value = r->GetRepeatedUInt32(message, field, index);
 			else
 				value = r->GetUInt32(message, field);
-			v = Number::New(value);
+			v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_UINT64: {
@@ -56,12 +54,12 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				uint32 hi, lo;
 				hi = (uint32) (value >> 32);
 				lo = (uint32) (value);
-				Local<Array> t = Array::New(2);
-				t->Set(0, Number::New(hi));
-				t->Set(1, Number::New(lo));
+				Local<Array> t = NanNew<Array>(2);
+				t->Set(0, NanNew<Number>(hi));
+				t->Set(1, NanNew<Number>(lo));
 				v = t;
 			} else
-				v = Number::New(value);
+				v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_DOUBLE: {
@@ -70,7 +68,7 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				value = r->GetRepeatedDouble(message, field, index);
 			else
 				value = r->GetDouble(message, field);
-			v = Number::New(value);
+			v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_FLOAT: {
@@ -79,7 +77,7 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				value = r->GetRepeatedFloat(message, field, index);
 			else
 				value = r->GetFloat(message, field);
-			v = Number::New(value);
+			v = NanNew<Number>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_BOOL: {
@@ -88,19 +86,19 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 				value = r->GetRepeatedBool(message, field, index);
 			else
 				value = r->GetBool(message, field);
-			v = Boolean::New(value);
+			v = NanNew<Boolean>(value);
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_ENUM: {
 			if (index >= 0)
-				v = String::New(r->GetRepeatedEnum(message, field, index)->name().c_str());
+				v = NanNew<String>(r->GetRepeatedEnum(message, field, index)->name().c_str());
 			else
-				v = String::New(r->GetEnum(message, field)->name().c_str());
+				v = NanNew<String>(r->GetEnum(message, field)->name().c_str());
 			break;
 		}
 		case FieldDescriptor::CPPTYPE_MESSAGE: {
 			if (field->is_optional() && !r->HasField(message, field))
-				v = Null();
+				v = NanNull();
 			else {
 				if (index >= 0)
 					v = ParsePart(r->GetRepeatedMessage(message, field, index));
@@ -116,20 +114,18 @@ Handle<Value> ParseField(const google::protobuf::Message &message, const Reflect
 			else
 				value = r->GetString(message, field);
 			if (field->type() == FieldDescriptor::TYPE_BYTES)
-				v = Buffer::New(const_cast<char *>(value.data()), value.length())->handle_;
+				v = NanNewBufferHandle(const_cast<char *>(value.data()), value.length());
 			else
-				v = String::New(value.c_str());
+				v = NanNew<String>(value.c_str());
 			break;
 		}
 	}
 
-	NanReturnValue(v);
+	return v;
 }
 
 Handle<Object> ParsePart(const google::protobuf::Message &message) {
-	NanScope();
-
-	Handle<Object> ret = Object::New();
+	Handle<Object> ret = NanNew<Object>();
 	// get a reflection
 	const Reflection *r = message.GetReflection();
 	const Descriptor *d = message.GetDescriptor();
@@ -144,7 +140,7 @@ Handle<Object> ParsePart(const google::protobuf::Message &message) {
 
 			if (field->is_repeated()) {
 				int size = r->FieldSize(message, field);
-				Handle<Array> array = Array::New(size);
+				Handle<Array> array = NanNew<Array>(size);
 				for (int i = 0; i < size; i++) {
 					array->Set(i, ParseField(message, r, field, i));
 				}
@@ -156,43 +152,9 @@ Handle<Object> ParsePart(const google::protobuf::Message &message) {
 			if (field->is_optional() && (v->IsNull() || !r->HasField(message, field)))
 				continue;
 			
-			ret->Set(String::NewSymbol(field->name().c_str()), v);
+			ret->Set(NanNew<String>(field->name().c_str()), v);
 		}
 	}
 
-	NanReturnValue(ret);
-}
-
-NAN_METHOD(Parse) {
-	NanScope();
-
-	Local<Value> p_pool = args[0]->ToObject()->GetInternalField(0);
-	DescriptorPool *pool = static_cast<DescriptorPool*>(External::Unwrap(p_pool));
-
-	Local<Object> buffer_obj = args[1]->ToObject();
-	char *buffer_data = Buffer::Data(buffer_obj);
-	size_t buffer_length = Buffer::Length(buffer_obj);
-
-	String::Utf8Value schemaName(args[2]->ToString());
-	std::string schema_name = std::string(*schemaName);
-
-	// create a message based on schema
-	DynamicMessageFactory factory;
-	const Descriptor *descriptor = pool->FindMessageTypeByName(schema_name);
-	if (descriptor == NULL) {
-		NanThrowError(("Unknown schema name: " + schema_name).c_str());
-		NanReturnNull();
-	}
-	google::protobuf::Message *message = factory.GetPrototype(descriptor)->New();
-
-	bool parseResult = message->ParseFromArray(buffer_data, buffer_length);
-
-	if (parseResult) {
-		Handle<Object> ret = ParsePart(*message);
-		delete message;
-		NanReturnValue(ret);
-	} else {
-		NanThrowError("Malformed protocol buffer");
-		NanReturnNull();
-	}
+	return ret;
 }